feat #22: Add Public Crate, Feature - Based Dependencies, and Separate Sinks/Sources
