+++ crates/fluxus-api/Cargo.toml
@@ -9,8 +9,12 @@ repository.workspace = true
 readme = "README.md"
 
 [dependencies]
-fluxus-core = "0.1"
-fluxus-runtime = "0.1"
+fluxus-core = { path = "../fluxus-core" }
+fluxus-runtime = { path = "../fluxus-runtime" }
+fluxus-utils = { path = "../fluxus-utils" }
+fluxus-sinks = { path = "../fluxus-sinks" }
+fluxus-sources = { path = "../fluxus-sources" }
+fluxus-transformers = { path = "../fluxus-transformers" }
 
 tokio = { version = "1", features = ["full"] }
 futures = "0.3"
