+++ crates/fluxus-core/src/pipeline/mod.rs
@@ -1,7 +1,5 @@
-mod dummy_sink;
 mod processor;
 mod status;
 
-pub use dummy_sink::DummySink;
 pub use processor::Pipeline;
 pub use status::PipelineStatus;
