+++ crates/fluxus-core/src/pipeline/processor.rs
@@ -1,4 +1,3 @@
-use super::DummySink;
 use super::status::PipelineStatus;
 use crate::BackpressureStrategy;
 use crate::Counter;
@@ -8,12 +7,13 @@ use crate::Timer;
 use crate::error_handling::BackpressureController;
 use crate::error_handling::ErrorHandler;
 use crate::metrics::Metrics;
-use crate::models::Record;
-use crate::models::StreamResult;
-use crate::operator::Operator;
-use crate::sink::Sink;
-use crate::source::Source;
-use crate::window::WindowConfig;
+use fluxus_sinks::Sink;
+use fluxus_sinks::dummy_sink::DummySink;
+use fluxus_sources::Source;
+use fluxus_transformers::operator::Operator;
+use fluxus_utils::models::Record;
+use fluxus_utils::models::StreamResult;
+use fluxus_utils::window::WindowConfig;
 use std::sync::Arc;
 use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
 use tokio::runtime::Handle;
