+++ crates/fluxus-runtime/Cargo.toml
@@ -9,7 +9,11 @@ repository.workspace = true
 readme = "README.md"
 
 [dependencies]
-fluxus-core = "0.1"
+fluxus-core = { path = "../fluxus-core" }
+fluxus-utils = { path = "../fluxus-utils" }
+fluxus-sinks = { path = "../fluxus-sinks" }
+fluxus-sources = { path = "../fluxus-sources" }
+fluxus-transformers = { path = "../fluxus-transformers" }
 
 tokio = { version = "1", features = ["full"] }
 futures = "0.3"
