+++ crates/fluxus-sources/src/csv.rs
@@ -1,5 +1,5 @@
-use crate::models::{Record, StreamResult};
 use async_trait::async_trait;
+use fluxus_utils::models::{Record, StreamResult};
 use std::path::PathBuf;
 use tokio::fs::File;
 use tokio::io::{AsyncBufReadExt, BufReader};
