+++ crates/fluxus-transformers/Cargo.toml
@@ -1,18 +1,20 @@
 [package]
-name = "fluxus-core"
-description = "Core components for Fluxus stream processing engine"
+name = "fluxus-transformers"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 authors.workspace = true
+description.workspace = true
+homepage.workspace = true
 repository.workspace = true
-readme = "README.md"
+readme.workspace = true
+categories.workspace = true
+keywords.workspace = true
 
 [dependencies]
 fluxus-utils = { path = "../fluxus-utils" }
 fluxus-sinks = { path = "../fluxus-sinks" }
 fluxus-sources = { path = "../fluxus-sources" }
-fluxus-transformers = { path = "../fluxus-transformers" }
 
 tokio = { version = "1", features = ["full"] }
 futures = "0.3"
