+++ crates/fluxus-transformers/src/transform_source.rs
@@ -1,5 +1,6 @@
 use async_trait::async_trait;
-use fluxus_core::{Record, Source, StreamResult};
+use fluxus_sources::Source;
+use fluxus_utils::models::{Record, StreamResult};
 use std::sync::Arc;
 
 use crate::{InnerOperator, InnerSource};
