+++ crates/fluxus/src/lib.rs
@@ -1,0 +1,34 @@
+#[cfg(feature = "fluxus-api")]
+pub mod api {
+    pub use fluxus_api::*;
+}
+
+#[cfg(feature = "fluxus-core")]
+pub mod core {
+    pub use fluxus_core::*;
+}
+
+#[cfg(feature = "fluxus-runtime")]
+pub mod runtime {
+    pub use fluxus_runtime::*;
+}
+
+#[cfg(feature = "fluxus-sinks")]
+pub mod sinks {
+    pub use fluxus_sinks::*;
+}
+
+#[cfg(feature = "fluxus-sources")]
+pub mod sources {
+    pub use fluxus_sources::*;
+}
+
+#[cfg(feature = "fluxus-transformers")]
+pub mod transformers {
+    pub use fluxus_transformers::*;
+}
+
+#[cfg(feature = "fluxus-utils")]
+pub mod utils {
+    pub use fluxus_utils::*;
+}
